Put all constants and default parameters into separate, organized files Slightly improved clarity of InputManager system Fixed timing issues relating to mixed types

+++ src/app/TimeManager.java
@@ -1,5 +1,3 @@
-
-
 package app;
 
 import java.util.HashMap;
@@ -31,14 +29,14 @@ public class TimeManager {
 
 	public static Map<Integer, Long> samples = new HashMap<Integer, Long>();
 
-	public static float lastSample(int hashCode) {
+	public static long lastSample(int hashCode) {
 		if (!samples.containsKey(hashCode)) {
 			return 0;
 		}
 		return samples.get(hashCode);
 	}
 
-	public static float sample(int hashCode) {
+	public static long sample(int hashCode) {
 		long time = time();
 		samples.put(hashCode, time);
 		return time;
